Throw exceptions if requested files don't exist.
Might need to change this later so that exceptions can optionally be silenced
to accomodate for ancestor chains where some of the ancestors don't exist.
For example, "dev.mariano" loads dev and dev.mariano, but the "dev" node might
not exist, and that might be OK (we only load the keys from "mariano").