Support PSR-0 autoloading, and Composer packaging.   (** BREAKS BACKWARDS COMPATABILITY **)
Added support for Composer[1] packaging so that this library
can be included more easily into other applications.

Also added support autoloading and namespaces in accordance
with PSR-0 [2], so that the library would be easier to use
and prevent name collisions. However, this caused the class
names to be renamed, which broke backwards compatability.

--
[1] PSR-0:    https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-0.md
[2] Composer: http://getcomposer.org/doc/02-libraries.md